pass function getCountryList to new user.js file location
