+++ client/src/actions/auth.js
@@ -45,7 +45,7 @@ export const logout = (username) => dispatch => {
 }
 
 export const countryList = (name) => {
-
-   const response = api.get(`/users/getCountryList?url_city=${name}`);  
+   const response = api.get(`/users/getCountryList?url_city=${name}`);
+   console.log(response);
    return response;
 }
